refactor: introduce prettier
