+++ lib/loader/loaderHelpers.js
@@ -3,5 +3,5 @@ const _ = require('lodash');
 const mergeConfigs = (configs) => configs.reduce((config, newC) => _.merge(config, newC), {});
 
 module.exports = {
-    mergeConfigs,
+  mergeConfigs,
 };
