+++ lib/processor/index.js
@@ -5,8 +5,8 @@ const mixin = require('./mixin');
 const replacer = require('./replacer');
 
 module.exports = (configToMergeInto, sourceConfigs, options) => {
-    const mergedSource = sourceConfigs.reduce((merged, source) => _.merge(merged, mixin(source, options)), {});
-    const mergeEnvConfig = selector(mergedSource, options);
-    const replaced = replacer(mergeEnvConfig, options);
-    return merger(configToMergeInto, replaced, options);
+  const mergedSource = sourceConfigs.reduce((merged, source) => _.merge(merged, mixin(source, options)), {});
+  const mergeEnvConfig = selector(mergedSource, options);
+  const replaced = replacer(mergeEnvConfig, options);
+  return merger(configToMergeInto, replaced, options);
 };
