+++ package.json
@@ -30,9 +30,12 @@
     "coveralls": "^3.0.3",
     "eslint": "7.27.0",
     "eslint-config-airbnb-base": "14.2.1",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-prettier": "3.4.0",
     "eslint-plugin-import": "2.23.4",
     "eslint-plugin-jest": "24.3.6",
-    "jest": "27.0.3"
+    "jest": "27.0.3",
+    "prettier": "2.3.0"
   },
   "author": "Doug Martin (doug.martin@c2fo.com)",
   "main": "index.js",
